requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.0
click==8.1.7
clickhouse-connect==0.6.21
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.8.0
duckdb==0.9.2
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.2.0
fastapi==0.99.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
graphlib-backport==1.0.3
h11==0.14.0
hnswlib==0.7.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
install==1.3.5
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.343
langchain-core==0.0.7
langdetect==1.0.9
langsmith==0.0.67
layoutparser==0.3.4
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.4
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.3.6
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.0.2
protobuf==4.25.1
pulsar-client==3.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.11
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
pypdfium2==4.24.0
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.12
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.11.0
unstructured-inference==0.7.16
unstructured.pytesseract==0.3.12
urllib3==1.26.6
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.3
zipp==3.17.0
zstandard==0.22.0
